Streamline existing code surrounding entity extraction.
