+++ bioagents/biosense/biosense_module.py
@@ -40,34 +40,18 @@ class BioSense_Module(Bioagent):
         context = content.get('context').to_string()
         # First get the KQML graph object for the given context
         graph = KQMLGraph(context)
-        # Then turn the graph into an EKB XML object, expanding around the
-        # given ID
-        ekb = EKB(graph, id_base)
-        ekb_str = ekb.to_string()
-        # Now process the EKB using the TRIPS processor to extract Statements
-        tp = trips.process_xml(ekb_str)
 
-        # If there are any statements then we can return the CL-JSON of those
-        if tp.statements:
-            cell_line_context = get_cell_line(ekb.ekb)
-            if cell_line_context:
-                set_cell_line_context(tp.statements, cell_line_context)
-            # Now make the CL-JSON for the given statements
-            js = self.make_cljson(tp.statements)
-        # Otherwise, we try extracting an Agent and return that
-        else:
-            try:
-                agent = agent_from_term(graph, id_base)
-                # Set the TRIPS ID in db_refs
-                agent.db_refs['TRIPS'] = id
-                # Infer the type from db_refs
-                agent = add_agent_type(agent)
-                js = self.make_cljson(agent)
-            except Exception as e:
-                logger.info("Encountered an error while parsing: %s. "
-                            "Returning empty list." % content.to_string())
-                logger.exception(e)
-                js = KQMLList()
+        try:
+            # Then turn the graph into an EKB XML object, expanding around the
+            # given ID
+            ekb = EKB(graph, id_base)
+            entity = ekb.get_entity()
+            js = self.make_cljson(entity)
+        except Exception as e:
+            logger.info("Encountered an error while parsing: %s. "
+                        "Returning empty list." % content.to_string())
+            logger.exception(e)
+            js = KQMLList()
         msg = KQMLPerformative('done')
         msg.sets('result', js)
         return msg
@@ -176,26 +160,6 @@ class BioSense_Module(Bioagent):
             msg = KQMLList('SUCCESS')
             msg.set('synonyms', syns_kqml)
         return msg
-
-
-# Look for a term representing a cell line
-def get_cell_line(et):
-    cl_tag = et.find("TERM/[type='ONT::CELL-LINE']/text")
-    if cl_tag is not None:
-        cell_line = cl_tag.text
-        cell_line.replace('-', '')
-        # TODO: add grounding here if available
-        clc = RefContext(cell_line)
-        return clc
-    return None
-
-
-def set_cell_line_context(stmts, context):
-    # Set cell line context if available
-    for stmt in stmts:
-        ev = stmt.evidence[0]
-        if not ev.context:
-            ev.context = BioContext(cell_line=context)
 
 
 if __name__ == "__main__":
